+++ Sources/PeripheryKit/Analyzer/Visitors/ObjCAccessibleRetainer.swift
@@ -26,6 +26,6 @@ final class ObjCAccessibleRetainer: SourceGraphVisitor {
             $0.attributes.contains("objc.name") ||
             $0.attributes.contains("objcMembers")
         }
-        objcDeclarations.forEach { $0.markRetained() }
+        objcDeclarations.forEach { graph.markRetained($0) }
     }
 }